Improve bash history SCT edge cases
